refactor: Component 파일이름 수정
