+++ src/components/atoms/SamplePost/SamplePost.jsx
