+++ src/components/index.js
@@ -7,23 +7,23 @@ export { Loading } from 'components/atoms/Loading/Loading';
 export { FormBtn } from 'components/atoms/FormBtn/FromBtn';
 export { FormInput } from 'components/atoms/FormInput/FormInput';
 export { FormLink } from 'components/atoms/FormLink/FormLink';
-export { Logo } from 'components/atoms/Logo';
-export { Label } from 'components/atoms/Label';
-export { Divider } from 'components/atoms/Divider';
-export { SamplePost } from 'components/atoms/SamplePost';
+export { Logo } from 'components/atoms/Logo/Logo';
+export { Label } from 'components/atoms/Label/Label';
+export { Divider } from 'components/atoms/Divider/Divider';
+export { SamplePost } from 'components/atoms/SamplePost/SamplePost';
 
 // molecules
 export { Form } from 'components/molecules/Form/Form';
 export { Profile } from 'components/molecules/Profile/Profile';
-export { HeaderButtons } from 'components/molecules/HeaderButtons';
-export { HeaderTexts } from 'components/molecules/HeaderTexts';
-export { FooterButtons } from 'components/molecules/FooterButtons';
-export { SamplePostList } from 'components/molecules/SamplePostList';
+export { HeaderButtons } from 'components/molecules/HeaderButtons/HeaderButtons';
+export { HeaderTexts } from 'components/molecules/HeaderTexts/HeaderTexts';
+export { FooterButtons } from 'components/molecules/FooterButtons/FooterButtons';
+export { SamplePostList } from 'components/molecules/SamplePostList/SamplePostList';
 
 // organisms
 export { ProfileList } from 'components/organisms/ProfileList/ProfileList';
-export { Header } from 'components/organisms/Header';
-export { Footer } from 'components/organisms/Footer';
+export { Header } from 'components/organisms/Header/Header';
+export { Footer } from 'components/organisms/Footer/Footer';
 
 // templates
-export { HomePage } from 'components/templates/HomePage';
+export { HomePage } from 'components/templates/HomePage/HomePage';
