+++ src/components/molecules/FooterButtons/FooterButtons.jsx
@@ -1,0 +1,12 @@
+import { Icon } from 'components';
+import * as S from './style';
+
+export function FooterButtons() {
+  return (
+    <S.FooterButtons>
+      <Icon icon="house" />
+      <Icon icon="calendar" />
+      <Icon icon="search" />
+    </S.FooterButtons>
+  );
+}
